# Speed up reduction stress test by eliminating repetitive array building