--- a/test/reductions/vass/reductions-stress-mf.chpl
+++ b/test/reductions/vass/reductions-stress-mf.chpl
@@ -6,10 +6,11 @@ config const r = 1000;
 writeln("n=", n, "   r=", r);
 var errors = 0;
 
+var d:domain(int);
+for i in 1..n {  d += i; }
+
 for rIdx in 1..r {
 
-  var d:domain(int);
-  for i in 1..n {  d += i; }
   var expected = n*(n+1)/2;
   var sumd = + reduce d;
   if (sumd != expected) {