@ronawho 2017-03-30
Note that this will "resolve" the reductions-stress-fast timeout under 32 bit, but David is still digging into the root cause of that regression.