@ronawho 2017-03-30
@vasslitvinov I'm going to merge to quiet the regression, but let me know if you have any objections post-commit